pyproject.toml
[tool.poetry]
name = "hikvision-api-wrapper"
version = "0.1.0"
description = ""
authors = ["loki"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
pytest = "^6.2.1"
vcrpy = "^4.1.1"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
